move routes into api folder
