+++ app/api/index.js
@@ -1,0 +1,3 @@
+const phrases = require("./phrases");
+
+module.exports = { phrases };
