+++ constants.js
@@ -1,0 +1,6 @@
+require("dotenv").config();
+
+module.exports = {
+  AIRTABLE_KEY: process.env.AIRTABLE_KEY,
+  AIRTABLE_BASE: process.env.AIRTABLE_BASE
+};
